+++ cpp/tests/unit-tests/StdComplexDataTypeTest.cpp
@@ -19,6 +19,7 @@
 #include <string>
 #include <unordered_set>
 
+#include <boost/math/special_functions/next.hpp>
 #include <gtest/gtest.h>
 
 #include "joynr/ByteBuffer.h"
@@ -443,4 +444,24 @@ TEST_F(StdComplexDataTypeTest, defaultVersionIsSetInStructInsideTypeCollectionWi
 
     EXPECT_EQ(expectedMajorVersion, joynr::types::TestTypesWithoutVersion::StructInsideTypeCollectionWithoutVersion::MAJOR_VERSION);
     EXPECT_EQ(expectedMinorVersion, joynr::types::TestTypesWithoutVersion::StructInsideTypeCollectionWithoutVersion::MINOR_VERSION);
+}
+
+
+TEST_F(StdComplexDataTypeTest, compareFloatingPointValues) {
+    using namespace boost::math;
+    TestTypes::TEverythingExtendedStruct struct1;
+    const float floatValue1 = 1.0f;
+    struct1.setTFloat(floatValue1);
+
+    // 3 floating point values above floatValue1
+    const float floatValue2 = float_next(float_next(float_next(floatValue1)));
+    TestTypes::TEverythingExtendedStruct struct2;
+    struct2.setTFloat(floatValue2);
+
+    // operator== compares 4 ULPs
+    EXPECT_TRUE(struct1 == struct2);
+
+    // use a custom precision
+    const std::size_t customMaxUlps = 2;
+    EXPECT_FALSE(struct1.equals(struct2, customMaxUlps));
 }